Intorduce command handler
