+++ WiredBrainCoffee ShopInfo Tool/ICommandHandler.cs
@@ -1,0 +1,7 @@
+﻿namespace WiredBrainCoffee_ShopInfo_Tool
+{
+    internal interface ICommandHandler
+    {
+        void HandleCommand();
+    }
+}
