+++ WiredBrainCoffee ShopInfo Tool/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 using WiredBrainCoffee.DataAccess;
 
 namespace WiredBrainCoffee_ShopInfo_Tool
@@ -20,32 +19,15 @@ namespace WiredBrainCoffee_ShopInfo_Tool
                 {
                     break;
                 }
-                if (string.Equals("help", line, StringComparison.OrdinalIgnoreCase))
-                {
-                    Console.WriteLine("> Available coffee shop commands:");
-                    foreach (var coffeeShop in coffeShops)
-                    {
-                        Console.WriteLine($"> " + coffeeShop.Location);
-                    }
-                }
-                else {
-                    var foundCoffeShops = coffeShops.Where(x => x.Location.StartsWith(line, StringComparison.OrdinalIgnoreCase)).ToList();
-                    if (foundCoffeShops.Count == 0)
-                        Console.WriteLine($"> Command {line} not found");
-                    else if (foundCoffeShops.Count == 1)
-                    {
-                        var coffeeShop = foundCoffeShops.Single();
-                        Console.WriteLine($"> Location: { coffeeShop.Location}");
-                        Console.WriteLine($"> Beans in stock: { coffeeShop.BeansInStokInKg} KG");
 
-                    }
-                    else Console.WriteLine("> Multiple matcching coffee shop commands found:");
-                    foreach (var coffeeShop in foundCoffeShops)
-                    {
-                        Console.WriteLine($"> " + coffeeShop.Location);
-                    }
-                }
+                var commandHandler = string.Equals("help", line, StringComparison.OrdinalIgnoreCase)
+                    ? new HelpCommandHandler(coffeShops) as ICommandHandler :
+                    new CoffeShopCommandHandler(line, coffeShops) as ICommandHandler;
+
+                commandHandler.HandleCommand();
             }
         }
+
+       
     }
 }
